Consolidate explore sections
