--- app/components/explore/sections/most-popular.js
@@ -1,25 +1,0 @@
-import BaseComponent from 'client/components/explore/sections/-base';
-import layout from 'client/templates/components/explore/sections/-base';
-import get from 'ember-metal/get';
-import { task } from 'ember-concurrency';
-
-export default BaseComponent.extend({
-  layout,
-  title: 'Most Popular',
-  more: 'explore.more',
-  name: 'most-popular',
-
-  init() {
-    this._super(...arguments);
-    get(this, 'getDataTask').perform().catch((error) => {
-      get(this, 'raven').captureException(error);
-    });
-  },
-
-  getDataTask: task(function* () {
-    return yield get(this, 'store').query(get(this, 'mediaType'), {
-      sort: '-userCount',
-      page: { limit: get(this, 'limit') }
-    });
-  }).restartable()
-});
