+++ app/templates/components/explore/explore-page.hbs
@@ -1,4 +1,4 @@
-<h3>{{t (concat "titles." mediaType ".index")}}</h3>
+<h3>{{title}}</h3>
 {{explore/category-widget}}
 {{explore/explore-search}}
 <div class="explore-content">
