--- app/templates/explore.hbs
@@ -1,3 +1,0 @@
-{{#explore/explore-page mediaType=mediaType}}
-  {{outlet}}
-{{/explore/explore-page}}
