+++ app/templates/explore/index.hbs
@@ -1,4 +1,26 @@
-{{explore/sections/top-airing mediaType=mediaType}}
-{{explore/sections/top-upcoming mediaType=mediaType}}
-{{explore/sections/highest-rated mediaType=mediaType}}
-{{explore/sections/most-popular mediaType=mediaType}}
+{{#explore/explore-page
+  title=(t (concat "titles." mediaType ".index"))
+  mediaType=mediaType}}
+  {{explore/explore-section
+    type="top-airing"
+    mediaType=mediaType
+    sort="-userCount"
+    filters="status:current"
+  }}
+  {{explore/explore-section
+    type="top-upcoming"
+    mediaType=mediaType
+    sort="-userCount"
+    filters="status:upcoming"
+  }}
+  {{explore/explore-section
+    type="highest-rated"
+    mediaType=mediaType
+    sort="-averageRating"
+  }}
+  {{explore/explore-section
+    type="most-popular"
+    mediaType=mediaType
+    sort="-userCount"
+  }}
+{{/explore/explore-page}}
