+++ translations/en-us.yaml
@@ -9,6 +9,12 @@ titles:
       episodes: "Episodes"
       characters: "Characters"
       reviews: "Reviews"
+  explore:
+    top-airing: "Top Airing {type}"
+    top-upcoming: "Top Upcoming {type}"
+    trending: "Trending {category} {type}"
+    highest-rated: "Highest Rated {type}"
+    most-popular: "Most Popular {type}"
   not-found: "Not Found"
   server-error: "Server Error"
   users:
